refactor(numerical-semigroups): restore wire models to _models.py
Move every private Pydantic request/result model and its validator
helpers back into the owner's _models.py per the architecture contract;
the split *_operations.py modules keep only execution helpers and run
functions, importing their contracts from _models.py.
