+++ src/jacobian/math/numerical_semigroups/_models.py
@@ -218,6 +218,96 @@ def _require_global_delta_bound(generators: tuple[int, ...]) -> None:
         )
 
 
+def _betti_component_index(
+    factorization: tuple[int, ...], components: tuple[tuple[int, ...], ...]
+) -> int:
+    """Return the unique Betti graph component supporting a factorization."""
+
+    support = {
+        index for index, coordinate in enumerate(factorization) if coordinate > 0
+    }
+    matches = [
+        index for index, component in enumerate(components) if support <= set(component)
+    ]
+    if len(matches) != 1:
+        raise _validation_error(
+            "relation factorization is not bound to one Betti component"
+        )
+    return matches[0]
+
+
+def _edges_span(component_count: int, edges: list[tuple[int, int]]) -> bool:
+    """Whether relation edges connect every component of one Betti element."""
+
+    reached = {0}
+    while True:
+        expanded = reached | {
+            right if left in reached else left
+            for left, right in edges
+            if left in reached or right in reached
+        }
+        if expanded == reached:
+            return reached == set(range(component_count))
+        reached = expanded
+
+
+def _require_exact_factorization_family(
+    generators: tuple[int, ...],
+    value: int,
+    family: tuple[tuple[int, ...], ...],
+) -> None:
+    """Replay one complete factorization family on its canonical axis."""
+
+    if len(set(family)) != len(family):
+        raise _validation_error("factorizations must be unique")
+    for factorization in family:
+        if len(factorization) != len(generators) or any(
+            coordinate < 0 for coordinate in factorization
+        ):
+            raise _validation_error("factorization has invalid coordinates")
+        degree = sum(
+            coordinate * generator
+            for coordinate, generator in zip(factorization, generators, strict=True)
+        )
+        if degree != value:
+            raise _validation_error(
+                "factorization does not evaluate to the result value"
+            )
+    if len(family) != factorization_count(generators, value):
+        raise _validation_error("factorizations do not form the complete family")
+
+
+def _factorization_graph_data(
+    family: tuple[tuple[int, ...], ...],
+) -> tuple[tuple[tuple[int, int], ...], tuple[tuple[int, ...], ...]]:
+    """Construct the canonical shared-support graph representation."""
+
+    edges = tuple(
+        (left, right)
+        for left in range(len(family))
+        for right in range(left + 1, len(family))
+        if any(
+            a > 0 and b > 0 for a, b in zip(family[left], family[right], strict=True)
+        )
+    )
+    unseen = set(range(len(family)))
+    components: list[tuple[int, ...]] = []
+    while unseen:
+        reached = {min(unseen)}
+        while True:
+            expanded = reached | {
+                right if left in reached else left
+                for left, right in edges
+                if left in reached or right in reached
+            }
+            if expanded == reached:
+                break
+            reached = expanded
+        unseen.difference_update(reached)
+        components.append(tuple(sorted(reached)))
+    return edges, tuple(components)
+
+
 class NumericalSemigroupRequest(StrictModel):
     """A numerical semigroup defined by a finite set of positive generators."""
 
@@ -592,6 +682,509 @@ class CatenaryDegreeResult(StrictModel):
         return self
 
 
+class NumericalSemigroupSummaryRequest(StrictModel):
+    """Compute the full summary of a numerical semigroup."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant; the summary uses "
+            "its increasing minimal generator axis."
+        ),
+    )
+
+    @model_validator(mode="after")
+    def require_positive_generators(self) -> Self:
+        _require_positive_bounded_generators(self.generators)
+        return self
+
+
+class NumericalSemigroupSummaryResult(StrictModel):
+    """Summary of a numerical semigroup."""
+
+    minimal_generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description="Increasing minimal generator axis of the numerical semigroup.",
+    )
+    multiplicity: CanonicalInteger
+    embedding_dimension: int = Field(ge=1)
+    frobenius_number: str
+    conductor: str
+    genus: int = Field(ge=0)
+    gaps: tuple[CanonicalInteger, ...]
+
+    @model_validator(mode="after")
+    def require_exact_summary(self) -> Self:
+        generators = _require_canonical_minimal_axis(self.minimal_generators)
+        (
+            multiplicity,
+            embedding_dimension,
+            frobenius_number,
+            conductor,
+            genus,
+            gaps,
+        ) = _summary_invariants(generators)
+        if parse_canonical_integer(self.multiplicity) != multiplicity:
+            raise _validation_error(
+                "multiplicity does not match the minimal generators"
+            )
+        if self.embedding_dimension != embedding_dimension:
+            raise _validation_error(
+                "embedding_dimension does not match the minimal generators"
+            )
+        if parse_canonical_integer(self.frobenius_number) != frobenius_number:
+            raise _validation_error(
+                "frobenius_number does not match the minimal generators"
+            )
+        if parse_canonical_integer(self.conductor) != conductor:
+            raise _validation_error("conductor does not match the minimal generators")
+        if self.genus != genus:
+            raise _validation_error("genus does not match the minimal generators")
+        if tuple(map(parse_canonical_integer, self.gaps)) != gaps:
+            raise _validation_error("gaps do not match the minimal generators")
+        return self
+
+
+class SemigroupMembershipRequest(StrictModel):
+    """Check membership of an integer in a numerical semigroup."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant."
+        ),
+    )
+    value: CanonicalInteger = Field(
+        description=f"Integer to test for membership, at most {MAX_ELEMENT}."
+    )
+
+    @model_validator(mode="after")
+    def require_positive_generators_and_bounded_value(self) -> Self:
+        _require_positive_bounded_generators(self.generators)
+        if parse_canonical_integer(self.value) > MAX_ELEMENT:
+            raise _validation_error(f"membership value must be at most {MAX_ELEMENT}")
+        return self
+
+
+class SemigroupMembershipResult(StrictModel):
+    """Whether the value is in the semigroup."""
+
+    value: CanonicalInteger
+    in_semigroup: bool
+
+
+class FactorizationComputeRequest(StrictModel):
+    """Compute the complete factorization family Z(s) for one element."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant; returned "
+            "coordinates use its increasing minimal generator axis."
+        ),
+    )
+    value: CanonicalInteger
+
+    @model_validator(mode="after")
+    def require_complete_materialization(self) -> Self:
+        generators = _require_minimal_generators(self.generators)
+        value = _require_bounded_value(self.value)
+        _require_materializable_factorizations(
+            generators, value, MAX_MATERIALIZED_FACTORIZATIONS
+        )
+        return self
+
+
+class FactorizationComputeResult(StrictModel):
+    """Complete factorization family Z(s) for one element."""
+
+    value: CanonicalInteger
+    minimal_generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1, max_length=MAX_GENERATORS
+    )
+    in_semigroup: bool
+    factorizations: tuple[tuple[int, ...], ...]
+
+    @model_validator(mode="after")
+    def require_exact_family(self) -> Self:
+        generators = _require_canonical_minimal_axis(self.minimal_generators)
+        value = parse_canonical_integer(self.value)
+        if self.in_semigroup != bool(self.factorizations):
+            raise _validation_error(
+                "membership must agree with the factorization family"
+            )
+        _require_exact_factorization_family(generators, value, self.factorizations)
+        return self
+
+
+class FactorizationLengthsComputeRequest(StrictModel):
+    """Compute the complete sorted length set of one element."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant; factorization "
+            "lengths use its increasing minimal generator axis."
+        ),
+    )
+    value: CanonicalInteger
+
+    @model_validator(mode="after")
+    def require_positive_generators_and_bounded_value(self) -> Self:
+        _require_minimal_generators(self.generators)
+        _require_bounded_value(self.value)
+        return self
+
+
+class FactorizationLengthsComputeResult(StrictModel):
+    """Sorted length set of one element."""
+
+    value: CanonicalInteger
+    minimal_generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1, max_length=MAX_GENERATORS
+    )
+    in_semigroup: bool
+    lengths: tuple[int, ...]
+
+    @model_validator(mode="after")
+    def require_length_set(self) -> Self:
+        generators = _require_canonical_minimal_axis(self.minimal_generators)
+        value = parse_canonical_integer(self.value)
+        if self.in_semigroup != bool(self.lengths):
+            raise _validation_error(
+                "membership must agree with the factorization lengths"
+            )
+        if self.lengths != tuple(sorted(set(self.lengths))):
+            raise _validation_error(
+                "lengths must be strictly increasing and duplicate-free"
+            )
+        if any(length < 0 for length in self.lengths):
+            raise _validation_error("factorization lengths must be non-negative")
+        if self.lengths != factorization_lengths(generators, value):
+            raise _validation_error("lengths do not form the complete length set")
+        return self
+
+
+class FactorizationDistanceRequest(StrictModel):
+    """Distance between two factorizations of the same element."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant, but both "
+            "factorization coordinate tuples must use its increasing minimal "
+            "generator axis."
+        ),
+    )
+    value: CanonicalInteger
+    first: tuple[int, ...] = Field(min_length=1)
+    second: tuple[int, ...] = Field(min_length=1)
+
+    @model_validator(mode="after")
+    def require_valid_request(self) -> Self:
+        generators = _require_minimal_generators(self.generators)
+        value = _require_bounded_value(self.value)
+        if any(c < 0 for c in self.first) or any(c < 0 for c in self.second):
+            raise _validation_error("factorization coordinates must be non-negative")
+        if len(self.first) != len(generators) or len(self.second) != len(generators):
+            raise _validation_error(
+                "factorization coordinates must match the minimal generating system"
+            )
+        first_value = sum(
+            coefficient * generator
+            for coefficient, generator in zip(self.first, generators, strict=True)
+        )
+        second_value = sum(
+            coefficient * generator
+            for coefficient, generator in zip(self.second, generators, strict=True)
+        )
+        if first_value != value or second_value != value:
+            raise _validation_error(
+                "both factorizations must evaluate to the declared value"
+            )
+        return self
+
+
+class FactorizationDistanceResult(StrictModel):
+    """Distance between two factorizations."""
+
+    value: CanonicalInteger
+    distance: int = Field(ge=0)
+    first_length: int = Field(ge=0)
+    second_length: int = Field(ge=0)
+
+
+class FactorizationGraphComputeRequest(StrictModel):
+    """Compute the standard factorization graph of one element."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant; graph vertices use "
+            "its increasing minimal generator axis."
+        ),
+    )
+    value: CanonicalInteger
+
+    @model_validator(mode="after")
+    def require_complete_materialization(self) -> Self:
+        generators = _require_minimal_generators(self.generators)
+        value = _require_bounded_value(self.value)
+        _require_materializable_factorizations(
+            generators, value, MAX_GRAPH_FACTORIZATIONS
+        )
+        return self
+
+
+class FactorizationGraphComputeResult(StrictModel):
+    """Standard factorization graph with connected components."""
+
+    value: CanonicalInteger
+    minimal_generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1, max_length=MAX_GENERATORS
+    )
+    in_semigroup: bool
+    factorizations: tuple[tuple[int, ...], ...]
+    edges: tuple[tuple[int, int], ...]
+    connected_components: tuple[tuple[int, ...], ...]
+    is_connected: bool
+
+    @model_validator(mode="after")
+    def require_graph_partition(self) -> Self:
+        generators = _require_canonical_minimal_axis(self.minimal_generators)
+        value = parse_canonical_integer(self.value)
+        vertex_count = len(self.factorizations)
+        if self.in_semigroup != bool(self.factorizations):
+            raise _validation_error("membership must agree with graph vertices")
+        vertices = tuple(
+            index for component in self.connected_components for index in component
+        )
+        if tuple(sorted(vertices)) != tuple(range(vertex_count)):
+            raise _validation_error(
+                "connected components must partition all graph vertices"
+            )
+        if self.is_connected != (len(self.connected_components) <= 1):
+            raise _validation_error("is_connected must agree with connected components")
+        for left, right in self.edges:
+            if not 0 <= left < right < vertex_count:
+                raise _validation_error("graph edge has invalid vertex indices")
+        _require_exact_factorization_family(generators, value, self.factorizations)
+        expected_edges, expected_components = _factorization_graph_data(
+            self.factorizations
+        )
+        if self.edges != expected_edges:
+            raise _validation_error("edges do not match shared-support adjacency")
+        if self.connected_components != expected_components:
+            raise _validation_error("connected components do not match the graph")
+        return self
+
+
+class MinimalPresentationRequest(StrictModel):
+    """One minimal presentation of a numerical semigroup."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant; returned relations "
+            "use its increasing minimal generator axis."
+        ),
+    )
+
+    @model_validator(mode="after")
+    def require_complete_candidate_range(self) -> Self:
+        generators = _require_minimal_generators(self.generators)
+        _require_global_betti_bound(generators)
+        return self
+
+
+class MinimalPresentationRelation(StrictModel):
+    """One relation (pair of distinct factorizations) in a presentation."""
+
+    first: tuple[int, ...]
+    second: tuple[int, ...]
+
+    @model_validator(mode="after")
+    def require_distinct_nonnegative_factorizations(self) -> Self:
+        if len(self.first) != len(self.second):
+            raise _validation_error("relation factorizations must have equal arity")
+        if any(value < 0 for value in (*self.first, *self.second)):
+            raise _validation_error("relation factorizations must be non-negative")
+        if self.first == self.second:
+            raise _validation_error("relation factorizations must be distinct")
+        return self
+
+
+class MinimalPresentationResult(StrictModel):
+    """One minimal presentation of the semigroup."""
+
+    minimal_generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1, max_length=MAX_GENERATORS
+    )
+    betti_elements: tuple[CanonicalInteger, ...]
+    relations: tuple[MinimalPresentationRelation, ...]
+
+    @model_validator(mode="after")
+    def require_minimal_relation_counts(self) -> Self:
+        generators = _require_canonical_minimal_axis(self.minimal_generators)
+        _, _, disconnected = betti_data(generators)
+        if tuple(map(parse_canonical_integer, self.betti_elements)) != tuple(
+            disconnected
+        ):
+            raise _validation_error(
+                "betti_elements do not match the minimal generators"
+            )
+        relation_components: dict[int, list[tuple[int, int]]] = {
+            betti: [] for betti in disconnected
+        }
+        for relation in self.relations:
+            if len(relation.first) != len(generators):
+                raise _validation_error(
+                    "relation arity does not match minimal generators"
+                )
+            first_degree = sum(
+                coordinate * generator
+                for coordinate, generator in zip(
+                    relation.first, generators, strict=True
+                )
+            )
+            second_degree = sum(
+                coordinate * generator
+                for coordinate, generator in zip(
+                    relation.second, generators, strict=True
+                )
+            )
+            if first_degree != second_degree or first_degree not in relation_components:
+                raise _validation_error("relation is not bound to a Betti element")
+            components = disconnected[first_degree]
+            left_component = _betti_component_index(relation.first, components)
+            right_component = _betti_component_index(relation.second, components)
+            if left_component == right_component:
+                raise _validation_error(
+                    "relation must connect distinct Betti components"
+                )
+            relation_components[first_degree].append((left_component, right_component))
+        expected = {
+            betti: len(components) - 1 for betti, components in disconnected.items()
+        }
+        if {
+            betti: len(edges) for betti, edges in relation_components.items()
+        } != expected:
+            raise _validation_error(
+                "relations do not have minimal per-Betti cardinality"
+            )
+        for betti, edges in relation_components.items():
+            if not _edges_span(len(disconnected[betti]), edges):
+                raise _validation_error("relations must span all Betti components")
+        return self
+
+
+class PresentationBinomialsRequest(StrictModel):
+    """Convert a minimal presentation to sparse binomial form."""
+
+    generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1,
+        max_length=MAX_GENERATORS,
+        description=(
+            f"Positive generators with gcd 1, each at most {MAX_GENERATOR}. "
+            "The presentation may be reordered or redundant; relation "
+            "coordinates must use its increasing minimal generator axis."
+        ),
+    )
+    relations: tuple[MinimalPresentationRelation, ...]
+
+    @model_validator(mode="after")
+    def require_kernel_relations(self) -> Self:
+        generators = _require_minimal_generators(self.generators)
+        for relation in self.relations:
+            if len(relation.first) != len(generators):
+                raise _validation_error(
+                    "relation coordinates must match the minimal generating system"
+                )
+            first_degree = sum(
+                coefficient * generator
+                for coefficient, generator in zip(
+                    relation.first, generators, strict=True
+                )
+            )
+            second_degree = sum(
+                coefficient * generator
+                for coefficient, generator in zip(
+                    relation.second, generators, strict=True
+                )
+            )
+            if first_degree != second_degree:
+                raise _validation_error(
+                    "relation factorizations must have the same semigroup degree"
+                )
+        return self
+
+
+class PresentationBinomial(StrictModel):
+    """One sparse binomial (aX - bX) arising from a presentation relation."""
+
+    left_coefficient: Literal["1"] = "1"
+    left_exponents: tuple[int, ...]
+    right_coefficient: Literal["-1"] = "-1"
+    right_exponents: tuple[int, ...]
+
+
+class PresentationBinomialsResult(StrictModel):
+    """Presentation converted to sparse binomials."""
+
+    minimal_generators: tuple[CanonicalInteger, ...] = Field(
+        min_length=1, max_length=MAX_GENERATORS
+    )
+    binomials: tuple[PresentationBinomial, ...]
+
+    @model_validator(mode="after")
+    def require_canonical_axis_and_homogeneous_binomials(self) -> Self:
+        generators = _require_canonical_minimal_axis(self.minimal_generators)
+        for binomial in self.binomials:
+            if len(binomial.left_exponents) != len(generators) or len(
+                binomial.right_exponents
+            ) != len(generators):
+                raise _validation_error(
+                    "binomial exponents must match the minimal generator axis"
+                )
+            if any(
+                exponent < 0
+                for exponent in (*binomial.left_exponents, *binomial.right_exponents)
+            ):
+                raise _validation_error("binomial exponents must be non-negative")
+            left_degree = sum(
+                exponent * generator
+                for exponent, generator in zip(
+                    binomial.left_exponents, generators, strict=True
+                )
+            )
+            right_degree = sum(
+                exponent * generator
+                for exponent, generator in zip(
+                    binomial.right_exponents, generators, strict=True
+                )
+            )
+            if left_degree != right_degree:
+                raise _validation_error(
+                    "binomial terms must have the same semigroup degree"
+                )
+        return self
+
+
 __all__ = [
     "BettiCatenaryDegree",
     "BettiElementsRequest",
@@ -608,4 +1201,22 @@ __all__ = [
     "ElementDeltaSetResult",
     "ElementElasticityRequest",
     "ElementElasticityResult",
+    "FactorizationComputeRequest",
+    "FactorizationComputeResult",
+    "FactorizationDistanceRequest",
+    "FactorizationDistanceResult",
+    "FactorizationGraphComputeRequest",
+    "FactorizationGraphComputeResult",
+    "FactorizationLengthsComputeRequest",
+    "FactorizationLengthsComputeResult",
+    "MinimalPresentationRelation",
+    "MinimalPresentationRequest",
+    "MinimalPresentationResult",
+    "NumericalSemigroupSummaryRequest",
+    "NumericalSemigroupSummaryResult",
+    "PresentationBinomial",
+    "PresentationBinomialsRequest",
+    "PresentationBinomialsResult",
+    "SemigroupMembershipRequest",
+    "SemigroupMembershipResult",
 ]
