+++ tests/math/numerical_semigroups/test_correctness_regressions.py
@@ -7,10 +7,6 @@ from math import lcm
 
 from tests.math.numerical_semigroups._support import numerical_semigroup_error
 
-from jacobian.math.numerical_semigroups._factorization_operations import (
-    FactorizationComputeRequest,
-    FactorizationGraphComputeRequest,
-)
 from jacobian.math.numerical_semigroups._models import (
     BettiElementsRequest,
     CatenaryDegreeRequest,
@@ -19,6 +15,11 @@ from jacobian.math.numerical_semigroups._models import (
     ElementCatenaryDegreeRequest,
     ElementDeltaSetRequest,
     ElementElasticityRequest,
+    FactorizationComputeRequest,
+    FactorizationGraphComputeRequest,
+    MinimalPresentationRequest,
+    MinimalPresentationResult,
+    PresentationBinomialsRequest,
 )
 from jacobian.math.numerical_semigroups._operations import (
     compute_betti_elements,
@@ -28,9 +29,6 @@ from jacobian.math.numerical_semigroups._operations import (
     compute_element_catenary_degree,
 )
 from jacobian.math.numerical_semigroups._presentation_operations import (
-    MinimalPresentationRequest,
-    MinimalPresentationResult,
-    PresentationBinomialsRequest,
     compute_minimal_presentation,
     compute_presentation_binomials,
 )
